feat(assets): add pagination filter, add sorting, add global providers in root layout
