+++ src/app/(onboarding)/@modal/(.)auth/sign-up/page.tsx
@@ -3,8 +3,6 @@
 import SignUpForm from "@/components/RegistrationForms/SignUpForm";
 import ErrorNotification from "@/components/UI/ErrorNotification/ErrorNotification";
 import RegistrationModal from "@/components/UI/RegistrationModal/RegistrationModal";
-import { CurrentUserProvider } from "@/stores/currentUser/CurrentUserProvider";
-import { TokenProvider } from "@/stores/tokenStore/TokenProvider";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
 
@@ -13,15 +11,11 @@ export default function SignUpModalPage() {
   const router = useRouter()
 
   const onClose = () => router.back()
-
+  
   return (
     <RegistrationModal isOpen onClose={onClose}>
       <ErrorNotification message={errorMessage} setErrorMessage={setErrorMessage} />
-      <CurrentUserProvider>
-        <TokenProvider>
-          <SignUpForm variant="modal" setErrorMessage={setErrorMessage}/>
-        </TokenProvider>
-      </CurrentUserProvider>
+      <SignUpForm variant="modal" setErrorMessage={setErrorMessage}/>
     </RegistrationModal>
   )
 }
