+++ src/app/auth/sign-up/page.tsx
@@ -5,8 +5,6 @@ import classes from "../auth.module.css"
 import SignUpForm from "@/components/RegistrationForms/SignUpForm";
 import { useState } from "react";
 import ErrorNotification from "@/components/UI/ErrorNotification/ErrorNotification";
-import { CurrentUserProvider } from "@/stores/currentUser/CurrentUserProvider";
-import { TokenProvider } from "@/stores/tokenStore/TokenProvider";
 
 export default function SignUpPage() {
   const [errorMessage, setErrorMessage] = useState("")
@@ -15,11 +13,7 @@ export default function SignUpPage() {
     <main className={classes.authPage}>
       <div className={classes.authCard}>
         <ErrorNotification message={errorMessage} setErrorMessage={setErrorMessage} />
-        <CurrentUserProvider>
-          <TokenProvider>
-            <SignUpForm variant="page" setErrorMessage={setErrorMessage} />
-          </TokenProvider>
-        </CurrentUserProvider>
+        <SignUpForm variant="page" setErrorMessage={setErrorMessage} />
       </div>
       <AbstractBackgroundShapes />
     </main>
