+++ src/components/UI/Assets/Assets.module.css
@@ -23,7 +23,7 @@
 
 .thead {
   position: relative;
-  z-index: 100;
+  z-index: 10;
 }
 
 .loadMoreSentinel {
@@ -105,29 +105,29 @@
 }
 
 .thName {
-  width: 15%;
+  width: 27.5%;
   position: sticky;
   left: var(--left-cell-inset);
 }
 
 .thPrice {
-  width: 7%;
+  width: 6.5%;
 }
 
 .thPct {
-  width: 6%;
+  width: 5%;
 }
 
 .thMCap {
-  width: 10%;
+  width: 8%;
 }
 
 .thFDV {
-  width: 10%;
+  width: 7%;
 }
 
 .thVolume {
-  width: 9%;
+  width: 7%;
 }
 
 .chartTh {
@@ -219,7 +219,7 @@
 }
 
 .assetIcon {
-  width: 20%;
+  width: 11%;
   aspect-ratio: 1 / 1;
   border-radius: 50%;
   background: var(--background-neutral-tinted-color); 
@@ -227,8 +227,10 @@
 
 .assetInfoFlex {
   display: flex;
+  width: 100%;
+  align-items: center;
   flex-wrap: nowrap;
-  gap: 8%;
+  gap: 0.6rem;
 }
 
 .assetName {
